+++ src/actions/miner.cpp
@@ -48,18 +48,19 @@ Miner::update ()
     if (!(slow_count % 3))
     {
       WorldObj::get_world()->remove(miner_radius,
-                                    static_cast<int>(pingu->get_x() - (miner_radius.get_width() / 2) + pingu->direction),
-                                    static_cast<int>(pingu->get_y() - miner_radius.get_width() + 1));
+                                    static_cast<int>(pingu->get_x() - static_cast<float>((miner_radius.get_width() / 2) + pingu->direction)),
+                                    static_cast<int>(pingu->get_y() - static_cast<float>(miner_radius.get_width() + 1)));
     }
 
-    pingu->set_pos(pingu->get_x() + pingu->direction, pingu->get_y() + 1);
+    pingu->set_pos(pingu->get_x() + static_cast<float>(pingu->direction),
+                   pingu->get_y() + 1.0f);
   }
 
   if (rel_getpixel(0, -1) == Groundtype::GP_NOTHING)
   {
     WorldObj::get_world()->remove(miner_radius,
-                                  static_cast<int>(pingu->get_x() - (miner_radius.get_width() / 2) + pingu->direction),
-                                  static_cast<int>(pingu->get_y() - miner_radius.get_width() + 3) );
+                                  static_cast<int>(pingu->get_x() - static_cast<float>((miner_radius.get_width() / 2) + pingu->direction)),
+                                  static_cast<int>(pingu->get_y() - static_cast<float>(miner_radius.get_width() + 3) ));
     pingu->set_action(Actions::WALKER);
   }
   else if (rel_getpixel(0, -1) == Groundtype::GP_SOLID
@@ -69,8 +70,8 @@ Miner::update ()
       Sound::PingusSound::play_sound("chink");
 
     WorldObj::get_world()->remove(miner_radius,
-                                  static_cast<int>(pingu->get_x() - (miner_radius.get_width() / 2) + pingu->direction),
-                                  static_cast<int>(pingu->get_y() - miner_radius.get_width() + 1) );
+                                  static_cast<int>(pingu->get_x() - static_cast<float>((miner_radius.get_width() / 2) + pingu->direction)),
+                                                   static_cast<int>(pingu->get_y() - static_cast<float>(miner_radius.get_width() + 1) ));
     pingu->set_action(Actions::WALKER);
 
     // Stop Pingu walking further into the solid.
