+++ src/editor/object_selector.cpp
@@ -236,7 +236,7 @@ public:
 
     if (mouse_over)
     {
-      int t_w = Fonts::verdana11.get_width(tooltip);
+      int t_w = static_cast<int>(Fonts::verdana11.get_width(tooltip));
       Rect t_r(rect.left + 17 - t_w/2 - 4, rect.top + 38 - 2, 
                rect.left + 17 + t_w/2 + 4, rect.top + 38 + Fonts::verdana11.get_height() + 4);
       gc.draw_fillrect(t_r, Color(255, 255, 200), 1000.0f);
